--- packages/local-app/src/types/LocalDocument.ts
@@ -1,8 +1,0 @@
-export { ProjectJSON } from "../../../node-data/src/project";
-
-export interface LocalDocument {
-  id: string;
-  name: string;
-  path: string;
-  updatedAt: number;
-}
